requirements.txt
asgiref==3.2.3
Django==3.0.3
django-registration-redux==2.7
django-widget-tweaks==1.4.5
pytz==2019.3
sqlparse==0.3.0
django-crispy-forms==1.9.2
requests==2.25.1
djangorestframework==3.12.2
django-oauth-toolkit
scipy==1.5.4
django-heroku
gunicorn
dj-database-url